+++ scaleph-support/scaleph-cache/src/main/java/cn/sliew/scaleph/cache/CaffeineCacheConfig.java
@@ -50,14 +50,4 @@ public class CaffeineCacheConfig {
         return caffeineCacheManager;
     }
 
-    public interface UnBoundedCaches {
-        String CACHE_DICT = "dict";
-        String CACHE_DICT_TYPE = "dictType";
-        String CACHE_ROLE = "role";
-        String CACHE_JOB_STEP_ATTR_TYPE = "jobStepAttrType";
-    }
-
-    public interface BoundedCaches {
-        String CACHE_DATA_SET = "dataSet";
-    }
 }
